adding pfp upload

+++ social-media-api/src/main/java/CIC/social_media_api/service/UserImageService.java
@@ -3,47 +3,52 @@ package CIC.social_media_api.service;
 import CIC.social_media_api.entity.User;
 import CIC.social_media_api.entity.UserImage;
 import CIC.social_media_api.repository.UserImageRepository;
+import CIC.social_media_api.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
-import java.util.Optional;
+import java.util.List;
 
 @Service
+@Transactional
 public class UserImageService {
 
     @Autowired
     private UserImageRepository userImageRepository;
 
     @Autowired
-    private UserService userService;
+    private UserRepository userRepository;
 
-    public UserImage createUserImage(Long userId, MultipartFile file) throws IOException {
-        Optional<User> optionalUser = userService.getUserById(userId);
+    public List<UserImage> getAllUserImages() {
+        return userImageRepository.findAll();
+    }
 
-        if (optionalUser.isPresent()) {
-            String fileName = file.getOriginalFilename();
-            String fileType = file.getContentType();
-            byte[] data = file.getBytes();
+    public UserImage getUserImageById(Long id) {
+        return userImageRepository.findById(id)
+                .orElseThrow(() -> new IllegalArgumentException("UserImage not found with ID: " + id));
+    }
 
-            UserImage userImage = new UserImage();
-            userImage.setName(fileName);
-            userImage.setType(fileType);
-            userImage.setData(data);
-            userImage.setUser(optionalUser.get());
+    public UserImage storeImage(MultipartFile file, Long userId) throws IOException {
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new IllegalArgumentException("User not found with ID: " + userId));
 
-            return userImageRepository.save(userImage);
-        } else {
-            throw new IllegalArgumentException("User not found with ID: " + userId);
+        UserImage userImage = new UserImage();
+        userImage.setName(file.getOriginalFilename());
+        userImage.setType(file.getContentType());
+        userImage.setData(file.getBytes());
+        userImage.setUser(user);
+
+        return userImageRepository.save(userImage);
+    }
+
+    public boolean deleteUserImageById(Long id) {
+        if (userImageRepository.existsById(id)) {
+            userImageRepository.deleteById(id);
+            return true;
         }
-    }
-
-    public Optional<UserImage> getUserImageByUserId(Long userId) {
-        return userImageRepository.findByUserId(userId); // Assuming findByUser_Id is correctly implemented in UserImageRepository
-    }
-
-    public void deleteUserImage(Long id) {
-        userImageRepository.deleteById(id);
+        return false;
     }
 }
